Refactor TorBalancerManager and remove complex TorPoolManager implementation
